If it's a unsigned long then the format string should be %lu.
This commit was SVN r5614.

+++ src/mca/pls/rsh/pls_rsh_module.c
@@ -266,7 +266,7 @@ int orte_pls_rsh_launch(orte_jobid_t jobid)
     }
 
     /* need integer value for command line parameter - NOT hex */
-    asprintf(&jobid_string, "%d", jobid);
+    asprintf(&jobid_string, "%lu", (unsigned long)jobid);
 
     /*
      * Build argv/env arrays.
